Recurse on directory redirects
When in recursive mode, if the response is a redirect to a directory, schedule
it to scan later.

The redirect needs to be resolved relative to the current URL to make relative
redirects work. We only add it to self.directories when it is a subdirectory,
to avoid endless loops when two subdirectories redirect to one another.